Refactor form editor screen
