+++ src/components/FormEditor/FormEditor.js
@@ -2,7 +2,6 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 import PropTypes from 'prop-types';
-import { get, map, toPairs, fromPairs } from 'lodash';
 import { ValidatedField, MultiSelect } from '../Form';
 import * as Fields from '../../ui/components/Fields';
 import SelectOptionImage from '../Form/Fields/SelectOptionImage';
@@ -12,106 +11,7 @@ import Disable from '../Disable';
 import NodeType from './NodeType';
 import { getFieldId } from '../../utils/issues';
 import { actionCreators as dialogsActions } from '../../ducks/modules/dialogs';
-
-
-const allowedTypes = ['text', 'number', 'boolean', 'ordinal', 'categorical'];
-
-const inputDefinitions = {
-  NumberInput: {
-    label: 'Number Input',
-    value: 'Number',
-    description: 'This input is optomized for collecting numerical data, and will show a number pad if available.',
-    image: 'TextInput',
-  },
-  Checkbox: {
-    label: 'Checkbox',
-    value: 'Checkbox',
-    description: 'This is a simple checkbox component that can be clicked or tapped to toggle a value between true or false.',
-    image: 'Checkbox',
-  },
-  CheckboxGroup: {
-    label: 'Checkbox Group',
-    value: 'CheckboxGroup',
-    description: 'This component provides a group of checkboxes so that multiple values can be toggled on or off.',
-    image: 'CheckboxGroup',
-  },
-  Toggle: {
-    label: 'Toggle',
-    value: 'Toggle',
-    description: 'This component renders a switch, which can be tapped or clicked to indicate "on" or "off".',
-    image: 'Toggle',
-  },
-  RadioGroup: {
-    label: 'Radio Group',
-    value: 'RadioGroup',
-    description: 'This will render a group of options and allow the user to choose one. Useful for likert-type scales or other ordinal variables.',
-    image: 'RadioGroup',
-  },
-  ToggleButton: {
-    label: 'Toggle Button',
-    value: 'ToggleButton',
-    description: 'This component provides a colorful button that can be toggled "on" or "off". It is useful for categorical variables where multiple options can be selected.',
-    image: 'ToggleButton',
-  },
-  ToggleButtonGroup: {
-    label: 'Toggle Button Group',
-    value: 'ToggleButtonGroup',
-    description: 'This component provides a colorful button that can be toggled "on" or "off". It is useful for categorical variables where multiple options can be selected.',
-    image: 'ToggleButtonGroup',
-  },
-  TextInput: {
-    label: 'Text Input',
-    value: 'Text',
-    description: 'This is a standard text input, allowing for simple data entry up to approximately 30 characters.',
-    image: 'TextInput',
-  },
-};
-
-const getInputsForType = (type) => {
-  switch (type) {
-    case 'number':
-      return [inputDefinitions.TextInput, inputDefinitions.NumberInput];
-    case 'text':
-      return [inputDefinitions.TextInput];
-    case 'boolean':
-      return [inputDefinitions.Checkbox, inputDefinitions.Toggle, inputDefinitions.ToggleButton];
-    case 'ordinal':
-      return [inputDefinitions.RadioGroup];
-    case 'categorical':
-      return [inputDefinitions.CheckboxGroup, inputDefinitions.ToggleButtonGroup];
-    default:
-      return [inputDefinitions.TextInput];
-  }
-};
-
-const optionGetter = (variables) => {
-  const allowedVariables = fromPairs(
-    toPairs(variables)
-      .filter(([, options]) => allowedTypes.includes(options.type)),
-  );
-  return (property, rowValues, allValues) => {
-    const variable = get(rowValues, 'variable');
-    switch (property) {
-      case 'variable': {
-        const used = map(allValues, 'variable');
-        return map(
-          allowedVariables,
-          (value, id) => ({
-            value: id,
-            label: value.label,
-            name: value.name,
-            description: value.description,
-            isDisabled: value !== variable && used.includes(value),
-          }),
-        );
-      }
-      case 'component':
-        return getInputsForType(get(variables, [variable, 'type']));
-      default:
-        return [{}];
-    }
-  };
-};
+import optionGetter from './optionGetter';
 
 class FormEditor extends Component {
   handleAttemptTypeChange = () => {
@@ -128,16 +28,10 @@ class FormEditor extends Component {
     const {
       nodeTypes,
       variables,
-      toggleCodeView,
     } = this.props;
 
     return (
       <div>
-        <div className="code-button">
-          <small>
-            (<a onClick={toggleCodeView}>Show Code View</a>)
-          </small>
-        </div>
         <h1 className="editor__heading">Edit Form</h1>
 
         <Guidance contentId="guidance.form.title">
@@ -204,7 +98,6 @@ class FormEditor extends Component {
 }
 
 FormEditor.propTypes = {
-  toggleCodeView: PropTypes.func.isRequired,
   resetFields: PropTypes.func.isRequired,
   nodeType: PropTypes.string,
   variables: PropTypes.object.isRequired,
